move git related logic to a seperate pkg
